+++ src/main/p2p/index.ts
@@ -1,10 +1,10 @@
-import {Orchestrator} from './Orchestrator'
-import {P2PCommand, P2PEvent} from './messages'
+import {SyncService} from './SyncService'
+import {P2PCommand, P2PEvent} from './types/messages'
 
 // Utility-process entry. Pure IPC adapter — every concern (chain.db,
-// peer pool, header/cfilter workers, status aggregation) lives in the
-// Orchestrator and below. This file exists only to bridge parentPort
-// messages to/from Orchestrator method calls.
+// peer pool, header/cfilter workers, status aggregation) lives in
+// SyncService and below. This file exists only to bridge parentPort
+// messages to/from SyncService method calls.
 
 declare const process: NodeJS.Process & {
   parentPort: {
@@ -13,7 +13,7 @@ declare const process: NodeJS.Process & {
   }
 }
 
-const orchestrator = new Orchestrator({
+const sync = new SyncService({
   status: status => process.parentPort.postMessage({type: 'status', status}),
   blockApplied: block => process.parentPort.postMessage({type: 'blockApplied', block}),
   cursorAdvanced: (walletId, height) =>
@@ -25,23 +25,23 @@ process.parentPort.on('message', ({data}) => {
   switch (data.type) {
     case 'start':
       console.log(data)
-      orchestrator.start(data).catch(err => {
+      sync.start(data).catch(err => {
         const message = err instanceof Error ? err.message : String(err)
         process.parentPort.postMessage({type: 'error', message})
       })
       return
     case 'stop':
       console.log(data)
-      orchestrator.stop().catch(err => {
+      sync.stop().catch(err => {
         const message = err instanceof Error ? err.message : String(err)
         process.parentPort.postMessage({type: 'error', message})
       })
       return
     case 'addWatchAddresses':
-      orchestrator.addWatchAddresses(data)
+      sync.addWatchAddresses(data)
       return
   }
 })
 
 // Push the initial 'idle' state to the parent.
-process.parentPort.postMessage({type: 'status', status: orchestrator.getStatus()})
+process.parentPort.postMessage({type: 'status', status: sync.getStatus()})
